requirements.txt
# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0

# Machine Learning
torch>=2.0.0
torchvision
torchaudio
transformers>=4.30.0
accelerate>=0.20.0

# Protein folding and analysis
esmfold
fair-esm
biopython
biotite
openmm
mdtraj

# Data science
numpy
pandas
matplotlib
seaborn
scikit-learn

# Experiment tracking
wandb
tensorboard

# Jupyter
jupyter
ipykernel
